requirements.txt
tqdm==4.64.1
Telethon==1.25.0
pandas==1.5.0
python-dotenv==0.21.0